refactor: extract no-op database callback
